New layout, extra settings activity, commentation

+++ app/src/main/java/de/till_s/spotifyshuffleradio/helper/spotify/tasks/SpotifyPlaylistTask.java
@@ -9,11 +9,11 @@ import com.spotify.sdk.android.player.Connectivity;
 
 import java.util.Map;
 
-import de.till_s.spotifyshuffleradio.Settings;
 import de.till_s.spotifyshuffleradio.helper.db.DbHelper;
 import de.till_s.spotifyshuffleradio.helper.db.DbHelperInsert;
 import de.till_s.spotifyshuffleradio.helper.db.DbUtils;
 import de.till_s.spotifyshuffleradio.helper.spotify.db.SpotifyPlaylistContract;
+import de.till_s.spotifyshuffleradio.utils.Settings;
 import de.till_s.spotifyshuffleradio.utils.Utils;
 import kaaes.spotify.webapi.android.SpotifyError;
 import kaaes.spotify.webapi.android.SpotifyService;
@@ -42,6 +42,14 @@ public class SpotifyPlaylistTask extends AsyncTask<Void, Void, Map<String, Strin
 
     private Callback<Map<String, String>> response;
 
+    /**
+     * Task for loading all user playlists from web API
+     *
+     * @param context        Android context
+     * @param forceReload    Force reload all playlists
+     * @param spotifyService The spotify web service from Spotify Web-API
+     * @param response       Callback
+     */
     public SpotifyPlaylistTask(Context context, boolean forceReload, SpotifyService spotifyService, Callback<Map<String, String>> response) {
         this.context = context;
         this.forceReload = forceReload;
@@ -53,15 +61,21 @@ public class SpotifyPlaylistTask extends AsyncTask<Void, Void, Map<String, Strin
         return DbUtils.getPlaylists(context);
     }
 
+    /**
+     * Refresh local database when forceReaload is set
+     *
+     * @param params ... Nothing
+     * @return Returns local saved playlists from database
+     */
     @Override
     protected Map<String, String> doInBackground(Void... params) {
         Settings.loadSettings(context);
 
-        if (((Settings.PLAYLIST_TIMESTAMP == -1 || (Utils.getTimestamp() - Settings.PLAYLIST_TIMESTAMP) > 24 * 60 * 60) || forceReload) && Utils.getNetworkConnectivity(context) != Connectivity.OFFLINE) {
+        if (forceReload && Utils.getNetworkConnectivity(context) != Connectivity.OFFLINE) {
             getApiPlaylists();
         }
 
-        return DbUtils.getPlaylists(context);
+        return getLocalPlaylists();
     }
 
     /**
@@ -76,39 +90,50 @@ public class SpotifyPlaylistTask extends AsyncTask<Void, Void, Map<String, Strin
         response.success(stringStringMap, null);
     }
 
+    /**
+     * Get spotify playlists from web api
+     */
     private void getApiPlaylists() {
         Pager<PlaylistSimple> playlistSimplePager = null;
 
-        try {
-            playlistSimplePager = spotifyService.getMyPlaylists();
-        } catch (RetrofitError error) {
-            SpotifyError spotifyError = SpotifyError.fromRetrofitError(error);
+        if (spotifyService != null) {
+            try {
+                playlistSimplePager = spotifyService.getMyPlaylists();
+            } catch (RetrofitError error) {
+                SpotifyError spotifyError = SpotifyError.fromRetrofitError(error);
 
-            Log.e(TAG, "Error: " + spotifyError.toString());
+                Log.e(TAG, "Error: " + spotifyError.toString());
+            }
+        } else {
+            Log.e(TAG, "Error ... spotify service null");
         }
 
 
-        DbUtils.deleteAllPlaylists(context);
+        if (playlistSimplePager != null) {
+            DbUtils.deleteAllPlaylists(context);
 
-        Log.i(TAG, "Count pager playlists: " + playlistSimplePager.items);
+            Log.i(TAG, "Count pager playlists: " + playlistSimplePager.items);
 
-        // Store playlists local
-        for (PlaylistSimple playlist : playlistSimplePager.items) {
-            String playlistID = playlist.id;
-            String playlistTitle = playlist.name;
+            // Store playlists local
+            for (PlaylistSimple playlist : playlistSimplePager.items) {
+                String playlistID = playlist.id;
+                String playlistTitle = playlist.name;
 
-            ContentValues dbPlaylist = new ContentValues();
+                ContentValues dbPlaylist = new ContentValues();
 
-            // Create new playlist entry in database
-            dbPlaylist.put(SpotifyPlaylistContract.SpoitfyPlaylistEntry.COLUMN_PLAYLIST_ID, playlistID);
-            dbPlaylist.put(SpotifyPlaylistContract.SpoitfyPlaylistEntry.COLUMN_NAME_TITLE, playlistTitle);
+                // Create new playlist entry in database
+                dbPlaylist.put(SpotifyPlaylistContract.SpoitfyPlaylistEntry.COLUMN_PLAYLIST_ID, playlistID);
+                dbPlaylist.put(SpotifyPlaylistContract.SpoitfyPlaylistEntry.COLUMN_NAME_TITLE, playlistTitle);
 
-            createPlaylistEntry(dbPlaylist);
+                createPlaylistEntry(dbPlaylist);
+            }
+
+            // Update playlist timestamp in settings
+            Settings.PLAYLIST_TIMESTAMP = Utils.getTimestamp();
+            Settings.saveSettings(context);
+        } else {
+            Log.e(TAG, "Error ... playlist pager null");
         }
-
-        // Update playlist timestamp in settings
-        Settings.PLAYLIST_TIMESTAMP = Utils.getTimestamp();
-        Settings.saveSettings(context);
     }
 
     /**
